Created keyring to store password encrypted
It was not easy create an automated auth for the keyring.
This was necessary in order to encrypt stored passwords and the likes.

Ended up finding a creative python script for the solution.

Guess that it is better to have some security,
even if it is by the means of obscurity.